Expose field resolver in mapping
